+++ sklearn/metrics/_plot/roc_curve.py
@@ -1,6 +1,6 @@
 from .. import auc
 from .. import roc_curve
-from ...utils._plot import _BinaryClassifierCurveDisplayMixin
+from ...utils._plotting import _BinaryClassifierCurveDisplayMixin
 
 
 class RocCurveDisplay(_BinaryClassifierCurveDisplayMixin):
@@ -101,7 +101,7 @@ class RocCurveDisplay(_BinaryClassifierCurveDisplayMixin):
         display : :class:`~sklearn.metrics.plot.RocCurveDisplay`
             Object that stores computed values.
         """
-        self.ax_, self.figure_, name = super()._validate_plot_params(ax=ax, name=name)
+        self.ax_, self.figure_, name = self._validate_plot_params(ax=ax, name=name)
 
         line_kwargs = {}
         if self.roc_auc is not None and name is not None:
@@ -214,7 +214,7 @@ class RocCurveDisplay(_BinaryClassifierCurveDisplayMixin):
         <...>
         >>> plt.show()
         """
-        y_pred, pos_label, name = super()._validate_and_get_response_values(
+        y_pred, pos_label, name = cls._validate_and_get_response_values(
             estimator,
             X,
             y,
@@ -316,7 +316,7 @@ class RocCurveDisplay(_BinaryClassifierCurveDisplayMixin):
         <...>
         >>> plt.show()
         """
-        pos_label_validated, name = super()._validate_from_predictions_params(
+        pos_label_validated, name = cls._validate_from_predictions_params(
             y_true, y_pred, sample_weight=sample_weight, pos_label=pos_label, name=name
         )
